services/ingest-worker/requirements.txt
fastapi==0.111.0
google-cloud-storage==2.17.0
psycopg2-binary==2.9.9
uvicorn[standard]==0.30.1
